+++ trixnity-clientserverapi/trixnity-clientserverapi-model/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/model/authentication/oauth2/responses/OAuth2ErrorResponse.kt
@@ -1,0 +1,6 @@
+package net.folivo.trixnity.clientserverapi.model.authentication.oauth2.responses
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class OAuth2ErrorResponse(val error: String, val description: String? = null)