+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for the OAuth2-based authentication API since Matrix v1.15 
+- Support for account management deep-linking (MSC4191)
 - Added inceptionYear based on LICENSE file and developer name tag for developers in POM to allow Gradle plugins
   like gradle-license-plugin or aboutLibraries to automatically fetch the correct Copyright information.
 - Add support for non suspend sha256 hashing